requirements.txt
# PyTorch and Deep Learning
torch>=1.8.0
torchvision>=0.9.0
torchaudio>=0.8.0
torchsummary>=1.5.1

# Scientific Computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine Learning
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Data Manipulation
einops>=0.3.0

# Serial Communication
pyserial>=3.5

# Optional Dependencies
fvcore>=0.1.5.post20220512

# Development and Testing (optional)
jupyter>=1.0.0
ipython>=7.0.0

# System and Utilities
tqdm>=4.62.0
datetime 